# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "UDNS WEB 后台服务，主要提供域名管理、线路管理、后台管理等接口功能。",
        "title": "UDNS-WEB api服务",
        "termsOfService": "http://test.udns.woa.com",
        "contact": {},
        "version": "1.0"
    },
    "host": "test.udns.woa.com",
    "basePath": "/api/v1/",
    "paths": {
        "/flow/status_notify": {
            "get": {
                "description": "QFlow状态推送",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "状态推送"
                ],
                "summary": "状态推送",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "推送ID",
                        "name": "_qflow_task_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/flow/task": {
            "post": {
                "description": "传递参数新增task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程管理"
                ],
                "summary": "新增流程任务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "新增task参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flow.StartTaskParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/flow/task/{task_id}": {
            "get": {
                "description": "根据taskID查询task详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程管理"
                ],
                "summary": "task详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task ID,流程ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/flow/task/{task_id}/steps/{step_id}/submit": {
            "put": {
                "description": "根据step ID提交步骤，通常在保存字段后操作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程管理"
                ],
                "summary": "提交步骤",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "step ID,步骤ID",
                        "name": "step_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task ID,任务ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/flow/task/{task_id}/stop": {
            "put": {
                "description": "根据taskID中止步骤，只有creator和管理员可以操作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程管理"
                ],
                "summary": "中止流程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "step ID,步骤ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/flow/tasks": {
            "post": {
                "description": "查询所有流程列表，需要管理员权限",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程列表"
                ],
                "summary": "流程列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询task参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flow.QueryTaskParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/flow/tasks/my_application": {
            "post": {
                "description": "查询我发起所有流程，支持各种参数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程列表"
                ],
                "summary": "我的申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询task参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flow.QueryTaskParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
        "/flow/tasks/my_approve": {
            "post": {
                "description": "查询由我审批的流程，包括当前待我审批，以及我审批完成的流程",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程列表"
                ],
                "summary": "我的审批",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "查询task参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flow.QueryTaskParams"
                        }
                    },
                    {
                        "type": "string",
                        "description": "当done=true查询已经审批完成的流程，否则查询当前待我审批的列表",
                        "name": "done",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
...
        "/flows": {
            "get": {
                "description": "查询所有流程类型，为发起流程查询流程提供参数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流程列表"
                ],
                "summary": "流程类型列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "StaffName",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/frame.AuthResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "flow.Basic": {
            "type": "object",
            "properties": {
                "child_item": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                }
            }
        },
        "flow.QueryTaskParams": {
            "type": "object",
            "properties": {
                "create_time_end": {
                    "type": "string"
                },
                "create_time_start": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "doing_step_owner": {
                    "type": "string"
                },
                "field_input_value": {
                    "type": "string"
                },
                "finish_time_end": {
                    "type": "string"
                },
                "finish_time_start": {
                    "type": "string"
                },
                "flow_id": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "status": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "step_owner": {
                    "type": "string"
                }
            }
        },
        "flow.StartTaskParams": {
            "type": "object",
            "properties": {
                "basic": {
                    "$ref": "#/definitions/flow.Basic"
                },
                "echo_fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "echo_step_owners": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "flow_id": {
                    "type": "integer"
                },
                "flow_type": {
                    "type": "string"
                }
            }
        },
        "frame.AuthResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
    }
}